@@ -1,26 +1,26 @@
 [package]
 name = "alacritty_config_derive"
-version = "0.2.2-dev"
+version = "0.2.3-dev"
 authors = ["Christian Duerr <contact@christianduerr.com>"]
-license = "MIT/Apache-2.0"
+license = "MIT OR Apache-2.0"
 description = "Failure resistant deserialization derive"
 homepage = "https://github.com/alacritty/alacritty"
 edition = "2021"
-rust-version = "1.65.0"
+rust-version = "1.70.0"
 
 [lib]
 proc-macro = true
 
 [dependencies]
-syn = { version = "2.0.16", features = ["derive", "parsing", "proc-macro", "printing"], default-features = false }
 proc-macro2 = "1.0.24"
 quote = "1.0.7"
+syn = { version = "2.0.16", features = ["derive", "parsing", "proc-macro", "printing"], default-features = false }
 
 [dev-dependencies.alacritty_config]
 path = "../alacritty_config"
-version = "0.1.2-dev"
+version = "0.2.1-dev"
 
 [dev-dependencies]
-serde = { version = "1.0.117", features = ["derive"] }
-serde_yaml = "0.8.14"
 log = "0.4.11"
+serde = { version = "1.0.117", features = ["derive"] }
+toml = "0.8.2"
