@@ -1,50 +1,44 @@
 [package]
 name = "alacritty_terminal"
-version = "0.19.1-dev"
+version = "0.22.1-dev"
 authors = ["Christian Duerr <contact@christianduerr.com>", "Joe Wilm <joe@jwilm.com>"]
 license = "Apache-2.0"
 description = "Library for writing terminal emulators"
 readme = "../README.md"
 homepage = "https://github.com/alacritty/alacritty"
 edition = "2021"
-rust-version = "1.65.0"
+rust-version = "1.70.0"
 
-[dependencies.alacritty_config_derive]
-path = "../alacritty_config_derive"
-version = "0.2.2-dev"
-
-[dependencies.alacritty_config]
-path = "../alacritty_config"
-version = "0.1.2-dev"
+[features]
+default = ["serde"]
+serde = ["dep:serde", "bitflags/serde", "vte/serde"]
 
 [dependencies]
-base64 = "0.13.0"
-bitflags = { version = "2.2.1", features = ["serde"] }
+base64 = "0.21.3"
+bitflags = "2.4.1"
 home = "0.5.5"
 libc = "0.2"
 log = "0.4"
-mio = "0.6.20"
-mio-extras = "2"
 parking_lot = "0.12.0"
-regex-automata = "0.1.9"
-smallvec = { version = "1.9.0", features = ["serde"] }
-serde = { version = "1", features = ["derive", "rc"] }
-serde_yaml = "0.8"
+polling = "3.0.0"
+regex-automata = "0.4.3"
 unicode-width = "0.1"
-vte = { version = "0.10.0", default-features = false }
+vte = { git = "https://github.com/chrhasse/vte-sixel.git", default-features = false, features = ["ansi", "serde"] }
+serde = { version = "1", features = ["derive", "rc"], optional = true }
+smallvec = { version = "1.13.1", features = ["serde"] }
 
 [target.'cfg(unix)'.dependencies]
-nix = { version = "0.26.2", default-features = false, features = ["term"] }
+rustix-openpty = "0.1.1"
 signal-hook = "0.3.10"
-signal-hook-mio = { version = "0.2.1", features = ["support-v0_6"] }
 
 [target.'cfg(windows)'.dependencies]
-mio-anonymous-pipes = "0.2"
-miow = "0.3"
-windows-sys = { version = "0.36", features = [
+piper = "0.2.1"
+miow = "0.6.0"
+windows-sys = { version = "0.48.0", features = [
     "Win32_System_Console",
     "Win32_Foundation",
     "Win32_Security",
+    "Win32_System_LibraryLoader",
     "Win32_System_Threading",
     "Win32_System_WindowsProgramming",
 ]}
