@@ -1,24 +1,26 @@
 use std::cmp::max;
-use std::os::raw::c_ulong;
+use std::ops::{Deref, DerefMut};
 use std::path::PathBuf;
+use std::rc::Rc;
 
-#[cfg(unix)]
-use clap::Subcommand;
-use clap::{ArgAction, Args, Parser, ValueHint};
-use log::{self, error, LevelFilter};
+use alacritty_config::SerdeReplace;
+use clap::{ArgAction, Args, Parser, Subcommand, ValueHint};
+use log::{error, LevelFilter};
 use serde::{Deserialize, Serialize};
-use serde_yaml::Value;
+use toml::Value;
 
-use alacritty_terminal::config::{Program, PtyConfig};
+use alacritty_terminal::tty::Options as PtyOptions;
 
+use crate::config::ui_config::Program;
 use crate::config::window::{Class, Identity};
-use crate::config::{serde_utils, UiConfig};
+use crate::config::UiConfig;
+use crate::logging::LOG_TARGET_IPC_CONFIG;
 
 /// CLI options for the main Alacritty executable.
 #[derive(Parser, Default, Debug)]
 #[clap(author, about, version = env!("VERSION"))]
 pub struct Options {
-    /// Print all events to stdout.
+    /// Print all events to STDOUT.
     #[clap(long)]
     pub print_events: bool,
 
@@ -30,17 +32,18 @@ pub struct Options {
     #[clap(long)]
     pub embed: Option<String>,
 
-    /// Specify alternative configuration file [default: $XDG_CONFIG_HOME/alacritty/alacritty.yml].
+    /// Specify alternative configuration file [default:
+    /// $XDG_CONFIG_HOME/alacritty/alacritty.toml].
     #[cfg(not(any(target_os = "macos", windows)))]
     #[clap(long, value_hint = ValueHint::FilePath)]
     pub config_file: Option<PathBuf>,
 
-    /// Specify alternative configuration file [default: %APPDATA%\alacritty\alacritty.yml].
+    /// Specify alternative configuration file [default: %APPDATA%\alacritty\alacritty.toml].
     #[cfg(windows)]
     #[clap(long, value_hint = ValueHint::FilePath)]
     pub config_file: Option<PathBuf>,
 
-    /// Specify alternative configuration file [default: $HOME/.config/alacritty/alacritty.yml].
+    /// Specify alternative configuration file [default: $HOME/.config/alacritty/alacritty.toml].
     #[cfg(target_os = "macos")]
     #[clap(long, value_hint = ValueHint::FilePath)]
     pub config_file: Option<PathBuf>,
@@ -58,20 +61,15 @@ pub struct Options {
     #[clap(short, conflicts_with("quiet"), action = ArgAction::Count)]
     verbose: u8,
 
-    /// Override configuration file options [example: cursor.style=Beam].
-    #[clap(short = 'o', long, num_args = 1..)]
-    option: Vec<String>,
-
     /// CLI options for config overrides.
     #[clap(skip)]
-    pub config_options: Value,
+    pub config_options: ParsedOptions,
 
     /// Options which can be passed via IPC.
     #[clap(flatten)]
     pub window_options: WindowOptions,
 
     /// Subcommand passed to the CLI.
-    #[cfg(unix)]
     #[clap(subcommand)]
     pub subcommands: Option<Subcommands>,
 }
@@ -80,14 +78,14 @@ impl Options {
     pub fn new() -> Self {
         let mut options = Self::parse();
 
-        // Convert `--option` flags into serde `Value`.
-        options.config_options = options_as_value(&options.option);
+        // Parse CLI config overrides.
+        options.config_options = options.window_options.config_overrides();
 
         options
     }
 
     /// Override configuration file with options from the CLI.
-    pub fn override_config(&self, config: &mut UiConfig) {
+    pub fn override_config(&mut self, config: &mut UiConfig) {
         #[cfg(unix)]
         {
             config.ipc_socket |= self.socket.is_some();
@@ -102,6 +100,9 @@ impl Options {
         if config.debug.print_events {
             config.debug.log_level = max(config.debug.log_level, LevelFilter::Info);
         }
+
+        // Replace CLI options.
+        self.config_options.override_config(config);
     }
 
     /// Logging filter level.
@@ -125,42 +126,6 @@ impl Options {
     }
 }
 
-/// Combine multiple options into a [`serde_yaml::Value`].
-pub fn options_as_value(options: &[String]) -> Value {
-    options.iter().fold(Value::default(), |value, option| match option_as_value(option) {
-        Ok(new_value) => serde_utils::merge(value, new_value),
-        Err(_) => {
-            eprintln!("Ignoring invalid option: {:?}", option);
-            value
-        },
-    })
-}
-
-/// Parse an option in the format of `parent.field=value` as a serde Value.
-fn option_as_value(option: &str) -> Result<Value, serde_yaml::Error> {
-    let mut yaml_text = String::with_capacity(option.len());
-    let mut closing_brackets = String::new();
-
-    for (i, c) in option.chars().enumerate() {
-        match c {
-            '=' => {
-                yaml_text.push_str(": ");
-                yaml_text.push_str(&option[i + 1..]);
-                break;
-            },
-            '.' => {
-                yaml_text.push_str(": {");
-                closing_brackets.push('}');
-            },
-            _ => yaml_text.push(c),
-        }
-    }
-
-    yaml_text += &closing_brackets;
-
-    serde_yaml::from_str(&yaml_text)
-}
-
 /// Parse the class CLI parameter.
 fn parse_class(input: &str) -> Result<Class, String> {
     let (general, instance) = match input.split_once(',') {
@@ -176,10 +141,10 @@ fn parse_class(input: &str) -> Result<Class, String> {
 }
 
 /// Convert to hex if possible, else decimal
-fn parse_hex_or_decimal(input: &str) -> Option<c_ulong> {
+fn parse_hex_or_decimal(input: &str) -> Option<u32> {
     input
         .strip_prefix("0x")
-        .and_then(|value| c_ulong::from_str_radix(value, 16).ok())
+        .and_then(|value| u32::from_str_radix(value, 16).ok())
         .or_else(|| input.parse().ok())
 }
 
@@ -206,8 +171,8 @@ impl TerminalOptions {
         Some(Program::WithArgs { program: program.clone(), args: args.to_vec() })
     }
 
-    /// Override the [`PtyConfig`]'s fields with the [`TerminalOptions`].
-    pub fn override_pty_config(&self, pty_config: &mut PtyConfig) {
+    /// Override the [`PtyOptions`]'s fields with the [`TerminalOptions`].
+    pub fn override_pty_config(&self, pty_config: &mut PtyOptions) {
         if let Some(working_directory) = &self.working_directory {
             if working_directory.is_dir() {
                 pty_config.working_directory = Some(working_directory.to_owned());
@@ -217,18 +182,18 @@ impl TerminalOptions {
         }
 
         if let Some(command) = self.command() {
-            pty_config.shell = Some(command);
+            pty_config.shell = Some(command.into());
         }
 
         pty_config.hold |= self.hold;
     }
 }
 
-impl From<TerminalOptions> for PtyConfig {
+impl From<TerminalOptions> for PtyOptions {
     fn from(mut options: TerminalOptions) -> Self {
-        PtyConfig {
+        PtyOptions {
             working_directory: options.working_directory.take(),
-            shell: options.command(),
+            shell: options.command().map(Into::into),
             hold: options.hold,
         }
     }
@@ -259,10 +224,11 @@ impl WindowIdentity {
 }
 
 /// Available CLI subcommands.
-#[cfg(unix)]
 #[derive(Subcommand, Debug)]
 pub enum Subcommands {
+    #[cfg(unix)]
     Msg(MessageOptions),
+    Migrate(MigrateOptions),
 }
 
 /// Send a message to the Alacritty socket.
@@ -289,6 +255,30 @@ pub enum SocketMessage {
     Config(IpcConfig),
 }
 
+/// Migrate the configuration file.
+#[derive(Args, Clone, Debug)]
+pub struct MigrateOptions {
+    /// Path to the configuration file.
+    #[clap(short, long, value_hint = ValueHint::FilePath)]
+    pub config_file: Option<PathBuf>,
+
+    /// Only output TOML config to STDOUT.
+    #[clap(short, long)]
+    pub dry_run: bool,
+
+    /// Do not recurse over imports.
+    #[clap(short = 'i', long)]
+    pub skip_imports: bool,
+
+    /// Do not move renamed fields to their new location.
+    #[clap(long)]
+    pub skip_renames: bool,
+
+    #[clap(short, long)]
+    /// Do not output to STDOUT.
+    pub silent: bool,
+}
+
 /// Subset of options that we pass to 'create-window' IPC subcommand.
 #[derive(Serialize, Deserialize, Args, Default, Clone, Debug, PartialEq, Eq)]
 pub struct WindowOptions {
@@ -299,13 +289,29 @@ pub struct WindowOptions {
     #[clap(flatten)]
     /// Window options which could be passed via IPC.
     pub window_identity: WindowIdentity,
+
+    #[clap(skip)]
+    #[cfg(target_os = "macos")]
+    /// The window tabbing identifier to use when building a window.
+    pub window_tabbing_id: Option<String>,
+
+    /// Override configuration file options [example: 'cursor.style="Beam"'].
+    #[clap(short = 'o', long, num_args = 1..)]
+    option: Vec<String>,
+}
+
+impl WindowOptions {
+    /// Get the parsed set of CLI config overrides.
+    pub fn config_overrides(&self) -> ParsedOptions {
+        ParsedOptions::from_options(&self.option)
+    }
 }
 
 /// Parameters to the `config` IPC subcommand.
 #[cfg(unix)]
 #[derive(Args, Serialize, Deserialize, Default, Debug, Clone, PartialEq, Eq)]
 pub struct IpcConfig {
-    /// Configuration file options [example: cursor.style=Beam].
+    /// Configuration file options [example: 'cursor.style="Beam"'].
     #[clap(required = true, value_name = "CONFIG_OPTIONS")]
     pub options: Vec<String>,
 
@@ -318,6 +324,78 @@ pub struct IpcConfig {
     /// Clear all runtime configuration changes.
     #[clap(short, long, conflicts_with = "options")]
     pub reset: bool,
+}
+
+/// Parsed CLI config overrides.
+#[derive(Debug, Default)]
+pub struct ParsedOptions {
+    config_options: Vec<(String, Value)>,
+}
+
+impl ParsedOptions {
+    /// Parse CLI config overrides.
+    pub fn from_options(options: &[String]) -> Self {
+        let mut config_options = Vec::new();
+
+        for option in options {
+            let parsed = match toml::from_str(option) {
+                Ok(parsed) => parsed,
+                Err(err) => {
+                    eprintln!("Ignoring invalid CLI option '{option}': {err}");
+                    continue;
+                },
+            };
+            config_options.push((option.clone(), parsed));
+        }
+
+        Self { config_options }
+    }
+
+    /// Apply CLI config overrides, removing broken ones.
+    pub fn override_config(&mut self, config: &mut UiConfig) {
+        let mut i = 0;
+        while i < self.config_options.len() {
+            let (option, parsed) = &self.config_options[i];
+            match config.replace(parsed.clone()) {
+                Err(err) => {
+                    error!(
+                        target: LOG_TARGET_IPC_CONFIG,
+                        "Unable to override option '{}': {}", option, err
+                    );
+                    self.config_options.swap_remove(i);
+                },
+                Ok(_) => i += 1,
+            }
+        }
+    }
+
+    /// Apply CLI config overrides to a CoW config.
+    pub fn override_config_rc(&mut self, config: Rc<UiConfig>) -> Rc<UiConfig> {
+        // Skip clone without write requirement.
+        if self.config_options.is_empty() {
+            return config;
+        }
+
+        // Override cloned config.
+        let mut config = (*config).clone();
+        self.override_config(&mut config);
+
+        Rc::new(config)
+    }
+}
+
+impl Deref for ParsedOptions {
+    type Target = Vec<(String, Value)>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.config_options
+    }
+}
+
+impl DerefMut for ParsedOptions {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.config_options
+    }
 }
 
 #[cfg(test)]
@@ -333,7 +411,7 @@ mod tests {
     use clap::CommandFactory;
     #[cfg(target_os = "linux")]
     use clap_complete::Shell;
-    use serde_yaml::mapping::Mapping;
+    use toml::Table;
 
     #[test]
     fn dynamic_title_ignoring_options_by_default() {
@@ -352,7 +430,7 @@ mod tests {
         let title = Some(String::from("foo"));
         let window_identity = WindowIdentity { title, ..WindowIdentity::default() };
         let new_window_options = WindowOptions { window_identity, ..WindowOptions::default() };
-        let options = Options { window_options: new_window_options, ..Options::default() };
+        let mut options = Options { window_options: new_window_options, ..Options::default() };
         options.override_config(&mut config);
 
         assert!(!config.window.dynamic_title);
@@ -371,38 +449,38 @@ mod tests {
     #[test]
     fn valid_option_as_value() {
         // Test with a single field.
-        let value = option_as_value("field=true").unwrap();
+        let value: Value = toml::from_str("field=true").unwrap();
 
-        let mut mapping = Mapping::new();
-        mapping.insert(Value::String(String::from("field")), Value::Bool(true));
+        let mut table = Table::new();
+        table.insert(String::from("field"), Value::Boolean(true));
 
-        assert_eq!(value, Value::Mapping(mapping));
+        assert_eq!(value, Value::Table(table));
 
         // Test with nested fields
-        let value = option_as_value("parent.field=true").unwrap();
+        let value: Value = toml::from_str("parent.field=true").unwrap();
 
-        let mut parent_mapping = Mapping::new();
-        parent_mapping.insert(Value::String(String::from("field")), Value::Bool(true));
-        let mut mapping = Mapping::new();
-        mapping.insert(Value::String(String::from("parent")), Value::Mapping(parent_mapping));
+        let mut parent_table = Table::new();
+        parent_table.insert(String::from("field"), Value::Boolean(true));
+        let mut table = Table::new();
+        table.insert(String::from("parent"), Value::Table(parent_table));
 
-        assert_eq!(value, Value::Mapping(mapping));
+        assert_eq!(value, Value::Table(table));
     }
 
     #[test]
     fn invalid_option_as_value() {
-        let value = option_as_value("}");
+        let value = toml::from_str::<Value>("}");
         assert!(value.is_err());
     }
 
     #[test]
     fn float_option_as_value() {
-        let value = option_as_value("float=3.4").unwrap();
+        let value: Value = toml::from_str("float=3.4").unwrap();
 
-        let mut expected = Mapping::new();
-        expected.insert(Value::String(String::from("float")), Value::Number(3.4.into()));
+        let mut expected = Table::new();
+        expected.insert(String::from("float"), Value::Float(3.4));
 
-        assert_eq!(value, Value::Mapping(expected));
+        assert_eq!(value, Value::Table(expected));
     }
 
     #[test]
