@@ -22,7 +22,7 @@ use std::cmp::max;
 use std::{fmt, mem};
 
 use crate::graphics::{ColorType, GraphicData, GraphicId, MAX_GRAPHIC_DIMENSIONS};
-use crate::term::color::Rgb;
+use crate::vte::ansi::Rgb;
 
 use log::trace;
 use vte::Params;
