@@ -2,10 +2,8 @@
 
 #![warn(rust_2018_idioms, future_incompatible)]
 #![deny(clippy::all, clippy::if_not_else, clippy::enum_glob_use)]
-#![cfg_attr(feature = "cargo-clippy", deny(warnings))]
+#![cfg_attr(clippy, deny(warnings))]
 
-pub mod ansi;
-pub mod config;
 pub mod event;
 pub mod event_loop;
 pub mod graphics;
@@ -20,3 +18,4 @@ pub mod vi_mode;
 
 pub use crate::grid::Grid;
 pub use crate::term::Term;
+pub use vte;
